libs: clear remaining clang warnings (graphics empty() sweep + 3 minor)
Finishes the warnings cleanup now that the graphics worktree WIP was
confirmed to be a stale duplicate (no real changes to conflict with).

- libs/graphics/src/library.cpp: migrate all 236 Value::empty()
  ([[deprecated]]) call sites to the equivalent Value() (Unset sentinel),
  same sweep as commit a7c215ad applied to core + the other libs.
- libs/stats/src/cluster/distance.cpp: add the missing
  Metric::Mahalanobis case to the row_distance switch (-Wswitch);
  Mahalanobis is computed via its own covariance path, never pairwise
  here, so it falls through to the existing default return.
- libs/image/src/io/tiff_writer.cpp: compute the FNV-1a hash in
  std::uint64_t and narrow to size_t on return, so the 64-bit
  offset-basis constant is not truncated where size_t is 32-bit
  (-Wconstant-conversion).
- multivariate.cpp / region.cpp / denoise.cpp: copy the structured
  binding into a plain local before capturing it in a lambda, so the
  capture is no longer a C++20 extension (-Wc++20-extensions).

Behaviour-preserving: full gtest 10317 passed = 10 baseline failures + 2
PCCTest.Canoncorr* RNG-flake (passed 5/5 on 3 re-runs), 0 new
regressions. Builds clean (0 errors/warnings on MSVC).

+++ libs/image/src/io/tiff_writer.cpp
@@ -148,12 +148,14 @@ private:
 
     struct VecHash {
         std::size_t operator()(const std::vector<std::uint8_t> &v) const {
-            std::size_t h = 1469598103934665603ull;
+            // 64-bit FNV-1a, computed in uint64 then narrowed to size_t so the
+            // offset-basis constant is not truncated where size_t is 32-bit.
+            std::uint64_t h = 1469598103934665603ull;
             for (auto b : v) {
                 h ^= b;
                 h *= 1099511628211ull;
             }
-            return h;
+            return static_cast<std::size_t>(h);
         }
     };
     std::unordered_map<std::vector<std::uint8_t>, std::uint32_t, VecHash> dict_;
